requirements.txt
Django>=4.2.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-dateutil>=2.8.2
Pillow>=10.0.0
qrcode>=7.4.2
PyJWT>=2.8.0
google-generativeai>=0.3.0
requests>=2.31.0

